:recycle: refactor: Renomeação de Keyboard para KeyboardChart - resolver erro de Single Name Component
